feat: remove client from db upon disconnect. might help #74
My understanding is that socketio maintains the mapping of room name and clients in the room in memory.
So this change is only relevant to the send_to_player function.

A drawback of this approach is that a client who is constantly disconnecting (because of bad internet), could potentially choke the database.