feat: Maintain scores separately. addresses #74
This new setup should have O(1) read and write times for scores. This is an improvement over the O(n) read time for scores, where n corresponds to the number of cards drawn. This was the cause for the increase in game latency.

This required a major change, so I expect scoring logic that was working before to break. Fixes to that will be added as we to test runs.


- add new score table
- integrate new score with the rest of the game